requirement.txt
Django==1.10.6
django-braces==1.11.0
django-cors-middleware==1.3.1
django-filter==1.0.1
django-oauth-toolkit==0.12.0
djangorestframework==3.6.2
Markdown==2.6.8
mysql-connector==2.1.4
MySQL-python==1.2.5
oauthlib==2.0.1
olefile==0.44
Pillow==4.1.1
python-dateutil==2.6.0
six==1.10.0
